MAINT: get rid of np.int
np.int is a confusing synonym for __builtin__.int; it maps to C long,
the size of which is platform-dependent. The NumPy devs are speaking of
deprecating np.int, and all uses of it in SciPy have been replaced.

Notably, this makes sample_without_replacement 64-bit clean; previously
it used np.int_t, which is a C long (32 bits wide on Windows).

See also http://github.com/numpy/numpy#6103.

+++ sklearn/cluster/k_means_.py
@@ -469,7 +469,7 @@ def _labels_inertia_precompute_dense(X, x_squared_norms, centers, distances):
 
     Returns
     -------
-    labels : numpy array, dtype=np.int, shape (n_samples,)
+    labels : numpy array, dtype=np.int32, shape (n_samples,)
         Indices of clusters that samples are assigned to.
 
     inertia : float
@@ -523,7 +523,7 @@ def _labels_inertia(X, x_squared_norms, centers,
 
     Returns
     -------
-    labels: int array of shape(n)
+    labels: int32 array of shape(n)
         The resulting assignment
 
     inertia : float
