+++ sklearn/preprocessing/label.py
@@ -507,7 +507,7 @@ def label_binarize(y, classes, neg_label=0, pos_label=1,
 
     if y_type == "binary":
         if len(classes) == 1:
-            Y = np.zeros((len(y), 1), dtype=np.int)
+            Y = np.zeros((len(y), 1), dtype=int)
             Y += neg_label
             return Y
         elif len(classes) >= 3:
@@ -637,12 +637,12 @@ def _inverse_binarize_thresholding(y, output_type, classes, threshold):
         if threshold > 0:
             if y.format not in ('csr', 'csc'):
                 y = y.tocsr()
-            y.data = np.array(y.data > threshold, dtype=np.int)
+            y.data = np.array(y.data > threshold, dtype=int)
             y.eliminate_zeros()
         else:
-            y = np.array(y.toarray() > threshold, dtype=np.int)
+            y = np.array(y.toarray() > threshold, dtype=int)
     else:
-        y = np.array(y > threshold, dtype=np.int)
+        y = np.array(y > threshold, dtype=int)
 
     # Inverse transform data
     if output_type == "binary":
@@ -734,7 +734,7 @@ class MultiLabelBinarizer(BaseEstimator, TransformerMixin):
             classes = sorted(set(itertools.chain.from_iterable(y)))
         else:
             classes = self.classes
-        dtype = np.int if all(isinstance(c, int) for c in classes) else object
+        dtype = int if all(isinstance(c, int) for c in classes) else object
         self.classes_ = np.empty(len(classes), dtype=dtype)
         self.classes_[:] = classes
         return self
@@ -767,7 +767,7 @@ class MultiLabelBinarizer(BaseEstimator, TransformerMixin):
         tmp = sorted(class_mapping, key=class_mapping.get)
 
         # (make safe for tuples)
-        dtype = np.int if all(isinstance(c, int) for c in tmp) else object
+        dtype = int if all(isinstance(c, int) for c in tmp) else object
         class_mapping = np.empty(len(tmp), dtype=dtype)
         class_mapping[:] = tmp
         self.classes_, inverse = np.unique(class_mapping, return_inverse=True)
